requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.109.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
packaging==23.2
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.0
python-telegram-bot==20.7
qbittorrent-api==2024.1.58
requests==2.31.0
sniffio==1.3.0
starlette==0.35.1
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.26.0
